+++ src/promptflow-devkit/promptflow/_sdk/_tracing.py
@@ -36,6 +36,7 @@ from promptflow._constants import (
     TraceEnvironmentVariableName,
 )
 from promptflow._sdk._constants import (
+    PF_SERVICE_HOST,
     PF_TRACE_CONTEXT,
     PF_TRACE_CONTEXT_ATTR,
     TRACE_DEFAULT_COLLECTION,
@@ -51,12 +52,12 @@ from promptflow._sdk._service.utils.utils import (
     is_port_in_use,
     is_run_from_built_binary,
 )
-from promptflow._sdk._tracing_utils import get_workspace_kind
-from promptflow._sdk._utils import (
+from promptflow._sdk._utilities.general_utils import (
     add_executable_script_to_env_path,
     extract_workspace_triad_from_trace_provider,
-    parse_kv_from_pb_attribute,
 )
+from promptflow._sdk._utilities.tracing_utils import get_workspace_kind, parse_kv_from_pb_attribute, parse_protobuf_span
+from promptflow._sdk.entities import Run
 from promptflow._utils.logger_utils import get_cli_sdk_logger
 from promptflow._utils.thread_utils import ThreadWithContextVars
 from promptflow.tracing._integrations._openai_injector import inject_openai_api
@@ -196,14 +197,15 @@ def _invoke_pf_svc() -> str:
     return port
 
 
-def _get_ws_triad_from_pf_config(path: typing.Optional[Path]) -> typing.Optional[AzureMLWorkspaceTriad]:
+def _get_ws_triad_from_pf_config(path: typing.Optional[Path], config=None) -> typing.Optional[AzureMLWorkspaceTriad]:
     from promptflow._sdk._configuration import Configuration
 
-    config = Configuration.get_instance().get_trace_destination(path=path)
-    _logger.info("resolved tracing.trace.destination: %s", config)
-    if not TraceDestinationConfig.need_to_export_to_azure(config):
+    config = config or Configuration.get_instance()
+    trace_destination = config.get_trace_destination(path=path)
+    _logger.info("resolved tracing.trace.destination: %s", trace_destination)
+    if not TraceDestinationConfig.need_to_export_to_azure(trace_destination):
         return None
-    return extract_workspace_triad_from_trace_provider(config)
+    return extract_workspace_triad_from_trace_provider(trace_destination)
 
 
 # priority: run > experiment > collection
@@ -225,7 +227,7 @@ def _get_tracing_url_from_local(
     exp: typing.Optional[str] = None,  # pylint: disable=unused-argument
     run: typing.Optional[str] = None,
 ) -> str:
-    url = f"http://localhost:{pfs_port}/v1.0/ui/traces/"
+    url = f"http://{PF_SERVICE_HOST}:{pfs_port}/v1.0/ui/traces/"
     if run is not None:
         url += f"?#run={run}"
     else:
@@ -334,7 +336,7 @@ def _inject_res_attrs_to_environ(
         os.environ[TraceEnvironmentVariableName.WORKSPACE_NAME] = ws_triad.workspace_name
     # we will not overwrite the value if it is already set
     if OTEL_EXPORTER_OTLP_ENDPOINT not in os.environ:
-        otlp_endpoint = f"http://localhost:{pfs_port}/v1/traces"
+        otlp_endpoint = f"http://{PF_SERVICE_HOST}:{pfs_port}/v1/traces"
         _logger.debug("set OTLP endpoint to environ: %s", otlp_endpoint)
         os.environ[OTEL_EXPORTER_OTLP_ENDPOINT] = otlp_endpoint
 
@@ -369,6 +371,11 @@ def start_trace_with_devkit(collection: str, **kwargs: typing.Any) -> None:
     _logger.debug("kwargs: %s", kwargs)
     attrs = kwargs.get("attributes", None)
     run = kwargs.get("run", None)
+    if isinstance(run, Run):
+        run_config = run._config
+        run = run.name
+    else:
+        run_config = None
     path = kwargs.get("path", None)
 
     # honor and set attributes if user has specified
@@ -396,7 +403,7 @@ def start_trace_with_devkit(collection: str, **kwargs: typing.Any) -> None:
 
     # local to cloud feature
     _logger.debug("start_trace_with_devkit.path(from kwargs): %s", path)
-    ws_triad = _get_ws_triad_from_pf_config(path=path)
+    ws_triad = _get_ws_triad_from_pf_config(path=path, config=run_config)
     is_azure_ext_installed = _is_azure_ext_installed()
     if ws_triad is not None and not is_azure_ext_installed:
         warning_msg = (
@@ -559,8 +566,8 @@ def process_otlp_trace_request(
     trace_request: ExportTraceServiceRequest,
     get_created_by_info_with_cache: typing.Callable,
     logger: logging.Logger,
+    get_credential: typing.Callable,
     cloud_trace_only: bool = False,
-    credential: typing.Optional[object] = None,
 ):
     """Process ExportTraceServiceRequest and write data to local/remote storage.
 
@@ -572,13 +579,12 @@ def process_otlp_trace_request(
     :type get_created_by_info_with_cache: Callable
     :param logger: The logger object used for logging.
     :type logger: logging.Logger
+    :param get_credential: A function that gets credential for Cosmos DB operation.
+    :type get_credential: Callable
     :param cloud_trace_only: If True, only write trace to cosmosdb and skip local trace. Default is False.
     :type cloud_trace_only: bool
-    :param credential: The credential object used to authenticate with cosmosdb. Default is None.
-    :type credential: Optional[object]
     """
     from promptflow._sdk.entities._trace import Span
-    from promptflow._sdk.operations._trace_operations import TraceOperations
 
     all_spans = []
     for resource_span in trace_request.resource_spans:
@@ -596,7 +602,7 @@ def process_otlp_trace_request(
         for scope_span in resource_span.scope_spans:
             for span in scope_span.spans:
                 # TODO: persist with batch
-                span: Span = TraceOperations._parse_protobuf_span(span, resource=resource, logger=logger)
+                span: Span = parse_protobuf_span(span, resource=resource, logger=logger)
                 if not cloud_trace_only:
                     all_spans.append(copy.deepcopy(span))
                     span._persist()
@@ -604,24 +610,20 @@ def process_otlp_trace_request(
                 else:
                     all_spans.append(span)
 
-    if cloud_trace_only:
-        # If we only trace to cloud, we should make sure the data writing is success before return.
-        _try_write_trace_to_cosmosdb(all_spans, get_created_by_info_with_cache, logger, credential, is_cloud_trace=True)
-    else:
-        # Create a new thread to write trace to cosmosdb to avoid blocking the main thread
-        ThreadWithContextVars(
-            target=_try_write_trace_to_cosmosdb,
-            args=(all_spans, get_created_by_info_with_cache, logger, credential, False),
-        ).start()
+    # Create a new thread to write trace to cosmosdb to avoid blocking the main thread
+    ThreadWithContextVars(
+        target=_try_write_trace_to_cosmosdb,
+        args=(all_spans, get_created_by_info_with_cache, logger, get_credential, cloud_trace_only),
+    ).start()
 
-    return
+    return all_spans
 
 
 def _try_write_trace_to_cosmosdb(
     all_spans: typing.List,
     get_created_by_info_with_cache: typing.Callable,
     logger: logging.Logger,
-    credential: typing.Optional[object] = None,
+    get_credential: typing.Callable,
     is_cloud_trace: bool = False,
 ):
     if not all_spans:
@@ -649,19 +651,31 @@ def _try_write_trace_to_cosmosdb(
         # So, we load clients in parallel for warm up.
         span_client_thread = ThreadWithContextVars(
             target=get_client,
-            args=(CosmosDBContainerName.SPAN, subscription_id, resource_group_name, workspace_name, credential),
+            args=(CosmosDBContainerName.SPAN, subscription_id, resource_group_name, workspace_name, get_credential),
         )
         span_client_thread.start()
 
         collection_client_thread = ThreadWithContextVars(
             target=get_client,
-            args=(CosmosDBContainerName.COLLECTION, subscription_id, resource_group_name, workspace_name, credential),
+            args=(
+                CosmosDBContainerName.COLLECTION,
+                subscription_id,
+                resource_group_name,
+                workspace_name,
+                get_credential,
+            ),
         )
         collection_client_thread.start()
 
         line_summary_client_thread = ThreadWithContextVars(
             target=get_client,
-            args=(CosmosDBContainerName.LINE_SUMMARY, subscription_id, resource_group_name, workspace_name, credential),
+            args=(
+                CosmosDBContainerName.LINE_SUMMARY,
+                subscription_id,
+                resource_group_name,
+                workspace_name,
+                get_credential,
+            ),
         )
         line_summary_client_thread.start()
 
@@ -677,7 +691,7 @@ def _try_write_trace_to_cosmosdb(
             subscription_id=subscription_id,
             resource_group_name=resource_group_name,
             workspace_name=workspace_name,
-            credential=credential,
+            get_credential=get_credential,
         )
 
         span_client_thread.join()
@@ -687,7 +701,7 @@ def _try_write_trace_to_cosmosdb(
 
         created_by = get_created_by_info_with_cache()
         collection_client = get_client(
-            CosmosDBContainerName.COLLECTION, subscription_id, resource_group_name, workspace_name, credential
+            CosmosDBContainerName.COLLECTION, subscription_id, resource_group_name, workspace_name, get_credential
         )
 
         collection_db = CollectionCosmosDB(first_span, is_cloud_trace, created_by)
@@ -697,31 +711,37 @@ def _try_write_trace_to_cosmosdb(
         # We assign it to LineSummary and Span and use it as partition key.
         collection_id = collection_db.collection_id
 
+        failed_span_count = 0
         for span in all_spans:
-            span_client = get_client(
-                CosmosDBContainerName.SPAN, subscription_id, resource_group_name, workspace_name, credential
-            )
-            result = SpanCosmosDB(span, collection_id, created_by).persist(
-                span_client, blob_container_client, blob_base_uri
-            )
-            # None means the span already exists, then we don't need to persist the summary also.
-            if result is not None:
-                line_summary_client = get_client(
-                    CosmosDBContainerName.LINE_SUMMARY,
-                    subscription_id,
-                    resource_group_name,
-                    workspace_name,
-                    credential,
+            try:
+                span_client = get_client(
+                    CosmosDBContainerName.SPAN, subscription_id, resource_group_name, workspace_name, get_credential
                 )
-                Summary(span, collection_id, created_by, logger).persist(line_summary_client)
-        collection_db.update_collection_updated_at_info(collection_client)
+                result = SpanCosmosDB(span, collection_id, created_by).persist(
+                    span_client, blob_container_client, blob_base_uri
+                )
+                # None means the span already exists, then we don't need to persist the summary also.
+                if result is not None:
+                    line_summary_client = get_client(
+                        CosmosDBContainerName.LINE_SUMMARY,
+                        subscription_id,
+                        resource_group_name,
+                        workspace_name,
+                        get_credential,
+                    )
+                    Summary(span, collection_id, created_by, logger).persist(line_summary_client)
+            except Exception as e:
+                failed_span_count += 1
+                stack_trace = traceback.format_exc()
+                logger.error(f"Failed to process span: {span.span_id}, error: {e}, stack trace: {stack_trace}")
+        if failed_span_count < len(all_spans):
+            collection_db.update_collection_updated_at_info(collection_client)
         logger.info(
             (
-                f"Finish writing trace to cosmosdb, total spans count: {len(all_spans)}."
-                f" Duration {datetime.now() - start_time}."
+                f"Finish writing trace to cosmosdb, total spans count: {len(all_spans)}, "
+                f"failed spans count: {failed_span_count}. Duration {datetime.now() - start_time}."
             )
         )
-
     except Exception as e:
         stack_trace = traceback.format_exc()
         logger.error(f"Failed to write trace to cosmosdb: {e}, stack trace is {stack_trace}")