+++ src/promptflow-evals/promptflow/evals/evaluators/_chat/__init__.py
@@ -1,0 +1,9 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+
+from ._chat import ChatEvaluator
+
+__all__ = [
+    "ChatEvaluator",
+]